Move blocking sidecar_interface to separate file
